pyproject.toml
[project]
name = "photo-heal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "piexif>=1.1.3",
    "pillow>=12.0.0",
]
